Island: Add build function for IAgentEventRepository
