+++ monkey/monkey_island/cc/services/__init__.py
@@ -13,3 +13,4 @@ from .agent_configuration_service import (
     register_resources as register_agent_configuration_resources,
 )  # noqa: E501
 from .agent_binary_service import IAgentBinaryService
+from .agent_binary_service import build as build_agent_binary_service
