+++ monkey/monkey_island/cc/services/agent_binary_service/__init__.py
@@ -1,1 +1,2 @@
 from .i_agent_binary_service import IAgentBinaryService
+from .build import build
