+++ monkey/monkey_island/cc/services/agent_binary_service/build.py
@@ -1,0 +1,66 @@
+import logging
+from pathlib import Path
+
+from common.utils.file_utils import get_binary_io_sha256_hash
+from monkey_island.cc.repositories import (
+    AgentBinaryRepository,
+    FileRepositoryCachingDecorator,
+    FileRepositoryLockingDecorator,
+    FileRepositoryLoggingDecorator,
+    IAgentBinaryRepository,
+    IFileRepository,
+    LocalStorageFileRepository,
+    RetrievalError,
+)
+from monkey_island.cc.server_utils.consts import MONKEY_ISLAND_ABS_PATH
+
+from . import IAgentBinaryService
+from .agent_binary_service import AgentBinaryService
+
+AGENT_BINARIES_PATH = Path(MONKEY_ISLAND_ABS_PATH) / "cc" / "binaries"
+logger = logging.getLogger(__name__)
+
+
+def build() -> IAgentBinaryService:
+    agent_binary_repository = _build_agent_binary_repository()
+
+    return AgentBinaryService(agent_binary_repository)
+
+
+def _build_agent_binary_repository() -> IAgentBinaryRepository:
+    file_repository = _decorate_file_repository(LocalStorageFileRepository(AGENT_BINARIES_PATH))
+    agent_binary_repository = AgentBinaryRepository(file_repository)
+
+    _log_agent_binary_hashes(agent_binary_repository)
+
+    return agent_binary_repository
+
+
+def _decorate_file_repository(file_repository: IFileRepository) -> IFileRepository:
+    return FileRepositoryLockingDecorator(
+        FileRepositoryLoggingDecorator(FileRepositoryCachingDecorator(file_repository))
+    )
+
+
+def _log_agent_binary_hashes(agent_binary_repository: IAgentBinaryRepository):
+    """
+    Logs all the hashes of the agent executables for debugging ease
+
+    :param agent_binary_repository: Used to retrieve the agent binaries
+    """
+    agent_binaries = {
+        "Linux": agent_binary_repository.get_linux_binary,
+        "Windows": agent_binary_repository.get_windows_binary,
+    }
+    agent_hashes = {}
+
+    for os, get_agent_binary in agent_binaries.items():
+        try:
+            agent_binary = get_agent_binary()
+            binary_sha256_hash = get_binary_io_sha256_hash(agent_binary)
+            agent_hashes[os] = binary_sha256_hash
+        except RetrievalError as err:
+            logger.error(f"No agent available for {os}: {err}")
+
+    for os, binary_sha256_hash in agent_hashes.items():
+        logger.info(f"{os} agent: SHA-256 hash: {binary_sha256_hash}")
